🧹 chore: Unifiy json and python serialization
