+++ src/ethereum_test_base_types/__init__.py
@@ -31,7 +31,7 @@ from .constants import (
 )
 from .conversions import to_bytes, to_hex
 from .json import to_json
-from .pydantic import CamelModel
+from .pydantic import CamelModel, EthereumTestBaseModel, EthereumTestRootModel
 from .reference_spec import ReferenceSpec
 
 __all__ = (
@@ -48,6 +48,8 @@ __all__ = (
     "EmptyOmmersRoot",
     "EmptyTrieRoot",
     "FixedSizeBytes",
+    "EthereumTestBaseModel",
+    "EthereumTestRootModel",
     "Hash",
     "HashInt",
     "HeaderNonce",
